Add atom, simplify renderable,
Add an Atom Renderer. One of the things The Room lost in the transition was an Atom Feed. This adds XML::Atom as a renderer for Blawd.

Simplify Renderer to always be an instance, but create a coercion to keep backcompat for a class name being passed in.

Move base_uri into the Renderer, and set it up so that we pass in an instance of the Renderer to the Atom and RSS indexes via Bread::Board.