The (first) big rendering refactor
The entries and indexes shouldn't be aware of how to render themselves. This is 90% of a re-factoring to make that true. Indexes still know how to render their entries, but I think that can be easily wiped out by making an index-specific renderer.

The rough gist of how we go about this is, we now have a Renderer factory that will pick the best choice of renderer class for a given `$entry`. The renderer is insansiated with global defaults that were set when the factory was created.